details view, order cart initial logic

+++ app/src/main/java/com/github/clockworkclyde/basedeliverymvvm/database/dao/OrderCartDao.kt
@@ -17,6 +17,6 @@ interface OrderCartDao {
     suspend fun insertOrderCartItem(item: OrderCartItem)
 
     @Query("DELETE FROM order_cart_items WHERE external_id LIKE :itemId")
-    suspend fun deleteOrderCartItem(itemId: Int)
+    suspend fun deleteOrderCartItem(itemId: Long)
 
 }
